Persistencia de Login, pegando o body da response e pequeno refatoramento
