+++ WebContent/indexLogin.xhtml
@@ -1,43 +1,40 @@
-<?xml version='1.0' encoding='UTF-8' ?>
-<!DOCTYPE html>
 <html xmlns="http://www.w3.org/1999/xhtml"
-	xmlns:ui="http://java.sun.com/jsf/facelets"
-	xmlns:f="http://java.sun.com/jsp/jstl/core"
 	xmlns:h="http://java.sun.com/jsf/html"
+	xmlns:f="http://java.sun.com/jsf/core"
 	xmlns:p="http://primefaces.org/ui">
 
 <h:head>
-	<meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-	<title>Login Page</title>
-
-	<link href="resources/css/login.css" rel="stylesheet" />
+	<title>Hatcher</title>
 </h:head>
-<h:body class="login-body">
 
-	<div class="login-wrapper">
-		<div class="login-panel">
-			<a href="home.xhtml" class="logo"><img
-				src="resources/images/hatcherEggLogo.png" /></a>
-			<h:form>
-				<p:inputText value="#{userBean2.user.login}" placeholder="Username"
-					class="username" />
-				<p:password value="#{userBean2.user.password}"
-					placeholder="Password" class="password" />
-				<h:commandButton action="#{userBean2.postLogin}" value="LOGIN"
-					class="loginButton" />
-			</h:form>
-			<a href="#">Esqueceu a senha?</a>
-			<p>
-				Não possui uma conta? <a href="signup.xhtml">Clique aqui para
-					cadastrar-se</a>
-			</p>
-		</div>
+<h:body>
+	<div>
+
+		<h:form>
+			<p:menubar>
+			
+				<p:submenu label="Usuarios" icon="pi pi-users" disabled="true">
+					<p:menuitem value="Novo Usuario" icon="pi pi-users" url="signup.xhtml" />
+					<p:menuitem value="Buscar Usuarios" icon="pi pi-search" url="userpage.xhtml" />
+				</p:submenu>
+
+				<p:submenu label="Projeto" icon="pi pi-briefcase" disabled="true">
+					<p:menuitem value="Novo projeto" url="#" />
+					<p:menuitem value="Buscar projeto" url="#" />
+				</p:submenu>
+
+				<p:submenu label="não sei" icon="pi pi-pencil">
+					<p:menuitem value="sei lá" url="#" icon="pi pi-angle-double-left" />
+					<p:menuitem value="sei lá" url="#" icon="pi pi-angle-double-right" />
+				</p:submenu>
+
+				<f:facet name="options">
+					<p:button type="button" value="Login" icon="pi pi-home"
+						href="login.xhtml" />
+				</f:facet>
+
+			</p:menubar>
+		</h:form>
 	</div>
-	<div class="login-footer">
-		<a href="www.ayty.org">Ayty.org</a>
-		<h6>Copyright 2021</h6>
-	</div>
-
-
 </h:body>
 </html>
