+++ WebContent/signup.xhtml
@@ -19,27 +19,25 @@
 				src="resources/images/hatcherEggLogo.png" /></a>
 			<h:form class="signupform">
 				<p>Nome completo</p>
-				<p:inputText value="#{userBean.user.fullName}"
+				<p:inputText value="#{userBean2.user.fullName}"
 					placeholder="Insira o nome completo" class="username" />
 
 				<p>Email</p>
-				<p:inputText value="#{userBean.user.email}"
+				<p:inputText value="#{userBean2.user.email}"
 					placeholder="Insira o seu Email" class="username" />
 				<p>Username</p>
-				<p:inputText value="#{userBean.user.login}" placeholder="Username"
+				<p:inputText value="#{userBean2.user.login}" placeholder="Username"
 					class="username" />
 
 				<p>Senha</p>
-				<p:password value="#{userBean.user.password}"
+				<p:password value="#{userBean2.user.password}"
 					placeholder="Password 1" class="password" />
-				<p>Digite novamente a senha</p>
-				<p:password placeholder="Password 2" class="password" />
 
 				<p>Foto de Perfil</p>
-				<p:inputText value="#{userBean.user.image}" placeholder="Image"
+				<p:inputText value="#{userBean2.user.image}" placeholder="Image"
 					class="username" />
 
-				<h:commandButton action="#{userBean.postUser}" value="CADASTRAR-SE"
+				<h:commandButton action="#{userBean2.postUser}" value="CADASTRAR-SE"
 					class="signupButton" />
 			</h:form>
