--- WebContent/successSignup.xhtml
@@ -1,12 +1,0 @@
-<?xml version='1.0' encoding='UTF-8' ?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-	xmlns:h="http://xmlns.jcp.org/jsf/html">
-<h:head>
-	<title>Success</title>
-</h:head>
-<h:body>
-	SUCCESS PAGE
-	<br />
-</h:body>
-</html>
