+++ src/main/java/org/ayty/bean/UserBean2.java
@@ -1,11 +1,13 @@
 package org.ayty.bean;
 
 import java.io.Serializable;
+import java.util.List;
 
 import javax.faces.view.ViewScoped;
 import javax.inject.Inject;
 import javax.inject.Named;
 
+import org.ayty.model.AccessToken;
 import org.ayty.model.User;
 import org.ayty.service.UserService2;
 
@@ -23,6 +25,8 @@ public class UserBean2 implements Serializable {
 	private String loginReturn;
 	private String username;
 
+	private AccessToken accessToken;
+
 	@Inject
 	UserService2 service;
 
@@ -39,12 +43,21 @@ public class UserBean2 implements Serializable {
 	}
 
 	public String postLogin() {
-		service.postLogin(user.getLogin(), user.getPassword());
+		accessToken = service.postLogin(user.getLogin(), user.getPassword());
 		loginReturn = service.login(user.getLogin(), user.getPassword());
 		this.setUsername(user.getLogin());
-		System.out.println("PostLogin user: " + getUsername());
 
 		return loginReturn;
+	}
+
+	public List<User> getUsers() {
+		return service.getUsers();
+
+	}
+
+	public String postUser() {
+		service.postUser(user.getLogin(), user.getPassword(), user.getEmail(), user.getFullName(), user.getImage());
+		return "indexLogout";
 	}
 
 }
