+++ src/main/java/org/ayty/model/AccessToken.java
@@ -1,0 +1,11 @@
+package org.ayty.model;
+
+import lombok.Data;
+
+@Data
+public class AccessToken {
+	
+	private String login;
+	private String token;
+
+}
