+++ src/main/java/org/ayty/service/UserService2.java
@@ -1,8 +1,8 @@
 package org.ayty.service;
 
 import java.io.Serializable;
+import java.util.List;
 
-import javax.faces.application.FacesMessage;
 import javax.faces.context.FacesContext;
 import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.client.Client;
@@ -10,10 +10,14 @@ import javax.ws.rs.client.ClientBuilder;
 import javax.ws.rs.client.Entity;
 import javax.ws.rs.client.WebTarget;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
+import javax.ws.rs.core.Response.Status;
 
+import org.ayty.model.AccessToken;
 import org.ayty.model.User;
 
 import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
 
 public class UserService2 implements Serializable {
 	private static final long serialVersionUID = 1L;
@@ -27,6 +31,8 @@ public class UserService2 implements Serializable {
 
 	User userLogin = new User();
 
+	AccessToken accessToken = new AccessToken();
+
 	public UserService2() {
 		this.facesContext = FacesContext.getCurrentInstance();
 		this.httpServletRequest = (HttpServletRequest) this.facesContext.getExternalContext().getRequest();
@@ -36,12 +42,13 @@ public class UserService2 implements Serializable {
 
 		userLogin.setLogin(username);
 		userLogin.setPassword(password);
-		if (userLogin.getLogin().isEmpty() && userLogin.getPassword().isEmpty()) {
+		System.out.println(userLogin.toString());
+		if (userLogin.getLogin().isEmpty() || userLogin.getPassword().isEmpty()) {
 			return "indexLogin";
 		} else {
 			this.httpServletRequest.getSession().setAttribute("sessionUsername", userLogin.getLogin());
 			System.out.println("Sessão aberta: " + userLogin.getLogin());
-			return "sucessoLogin";
+			return "indexLogout";
 		}
 
 	}
@@ -62,12 +69,64 @@ public class UserService2 implements Serializable {
 		return userLoginJson;
 	}
 
-	public void postLogin(String username, String password) {
+	public AccessToken postLogin(String username, String password) {
 		String userLoginJson = toJsonLogin(username, password);
 		Client client = ClientBuilder.newClient();
 		WebTarget target = client.target(REST_URI_POST_LOGIN);
-		client.target(REST_URI_POST_LOGIN).request(MediaType.APPLICATION_JSON)
+
+		Response response = client.target(REST_URI_POST_LOGIN).request(MediaType.APPLICATION_JSON)
 				.post(Entity.entity(userLoginJson, MediaType.APPLICATION_JSON));
+
+		if (response.getStatus() == Status.UNAUTHORIZED.getStatusCode()) {
+			return accessToken = new Gson().fromJson(response.readEntity(String.class), AccessToken.class);
+		} else {
+			return accessToken = new Gson().fromJson(response.readEntity(String.class), AccessToken.class);
+		}
+	}
+
+	public List<User> getUsers() {
+
+		Client client = ClientBuilder.newClient();
+
+		String URL = REST_URI_GET_USER;
+
+		WebTarget target = client.target(URL);
+
+		Response response = target.request().get();
+
+		String json = response.readEntity(String.class);
+		response.close();
+
+		List<User> list = new Gson().fromJson(json, new TypeToken<List<User>>() {
+		}.getType());
+		return list;
+	}
+
+	private String toJson(String username, String password, String email, String fullname, String image) {
+
+		User user = new User();
+		user.setProfile("ALUNO");
+		user.setAdmin(false);
+		user.setEmail(email);
+		user.setFullName(fullname);
+		user.setImage(image);
+		user.setPassword(password);
+		user.setLogin(username);
+
+		String userRegisterJson = new Gson().toJson(user);
+
+		return userRegisterJson;
+	}
+
+	public void postUser(String username, String password, String email, String fullname, String image) {
+		String userRegisterJson = toJson(username, password, email, fullname, image);
+		Client client = ClientBuilder.newClient();
+		WebTarget target = client.target(REST_URI_POST_USER);
+//		client.target(REST_URI_POST_USER).request(MediaType.APPLICATION_JSON)
+//				.post(Entity.entity(userRegisterJson, MediaType.APPLICATION_JSON));
+
+		Response response = client.target(REST_URI_POST_USER).request(MediaType.APPLICATION_JSON)
+				.post(Entity.entity(userRegisterJson, MediaType.APPLICATION_JSON));
 
 	}
